requirements.txt
streamlit 
google-generativeai
python-dotenv
mysql-connector-python